Convert the existing waiver to LaTeX
The old waiver was generated in Microsoft Word. It was difficult to tell
which affiliation you were actually selecting, and the options didn't
exactly match what we actually collect in the Trips and gear database.

This new form has nicer typography, more room to complete responses, and
a simpler interface for specifying affiliation.