Update to new template JSON
The PDF is exactly the same, but we use newer JSON constructs.
By running on Python 3.7, dictionaries are written to JSON in the exact
same order they were parsed (which makes diffing considerably simpler!)

Rotionale for creating a new template:
======================================

I noticed that the "Numbers" validation on phone numbers no longer
allows the E.164 format (more specifically, it does not like the leading
plus symbol for country code). This leads to the pre-filled phone
numbers no longer being accepted! This is less than great UX.

Because phone numbers should be able to have other characters too (like
parens, or the `x` character for an extension), we should just make the
field free-form text.

It's been a while since the last JSON update, so this template JSON
includes a number of DocuSign updates along with it. The only
intentional change made before re-exporting though was simply changing
the validation on two phone number fields.